* retrofitted some of the old KeyState code back into the tdns library.   It is not completely integrated yet, but the code compiles.
